# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sovereign Headless Backend Platform",
    "description": "High-performance async micro-kernel backend platform for sovereign enterprise apps.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/mail_gateway/servers": {
      "get": {
        "tags": [
          "Mail Gateway & Async Email Pipeline",
          "Mail Gateway - Servers"
        ],
        "summary": "List tenant SMTP outbound mail servers",
        "description": "Retrieve all configured outbound SMTP servers for the authenticated tenant.",
        "operationId": "list_mail_servers_api_v1_mail_gateway_servers_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by active status",
              "title": "Is Active"
            },
            "description": "Filter by active status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MailServerRead"
                  },
                  "title": "Response List Mail Servers Api V1 Mail Gateway Servers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Mail Gateway & Async Email Pipeline",
          "Mail Gateway - Servers"
        ],
        "summary": "Create SMTP server configuration",
        "description": "Register a new outbound SMTP server for the active company.",
        "operationId": "create_mail_server_api_v1_mail_gateway_servers_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailServerCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailServerRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mail_gateway/servers/{server_id}": {
      "get": {
        "tags": [
          "Mail Gateway & Async Email Pipeline",
          "Mail Gateway - Servers"
        ],
        "summary": "Get SMTP server details",
        "description": "Retrieve details for a specific outbound SMTP server configuration.",
        "operationId": "get_mail_server_api_v1_mail_gateway_servers__server_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Server Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailServerRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Mail Gateway & Async Email Pipeline",
          "Mail Gateway - Servers"
        ],
        "summary": "Update SMTP server configuration",
        "description": "Update fields on an existing outbound SMTP server configuration.",
        "operationId": "update_mail_server_api_v1_mail_gateway_servers__server_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Server Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailServerUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailServerRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Mail Gateway & Async Email Pipeline",
          "Mail Gateway - Servers"
        ],
        "summary": "Soft-delete SMTP server configuration",
        "description": "Soft delete an outbound SMTP configuration.",
        "operationId": "delete_mail_server_api_v1_mail_gateway_servers__server_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Server Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mail_gateway/servers/{server_id}/test": {
      "post": {
        "tags": [
          "Mail Gateway & Async Email Pipeline",
          "Mail Gateway - Servers"
        ],
        "summary": "Test SMTP connection and authentication",
        "description": "Test network handshake, TLS negotiation, and credentials for an SMTP configuration.",
        "operationId": "test_mail_server_connection_api_v1_mail_gateway_servers__server_id__test_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Server Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mail_gateway/templates": {
      "get": {
        "tags": [
          "Mail Gateway & Async Email Pipeline",
          "Mail Gateway - Templates"
        ],
        "summary": "List email templates",
        "description": "List Jinja2 email templates for the authenticated tenant.",
        "operationId": "list_mail_templates_api_v1_mail_gateway_templates_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "model_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by target model name",
              "title": "Model Name"
            },
            "description": "Filter by target model name"
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by language code",
              "title": "Language"
            },
            "description": "Filter by language code"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MailTemplateRead"
                  },
                  "title": "Response List Mail Templates Api V1 Mail Gateway Templates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Mail Gateway & Async Email Pipeline",
          "Mail Gateway - Templates"
        ],
        "summary": "Create email template",
        "description": "Register a new parameterized Jinja2 email template.",
        "operationId": "create_mail_template_api_v1_mail_gateway_templates_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailTemplateCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailTemplateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mail_gateway/templates/render": {
      "post": {
        "tags": [
          "Mail Gateway & Async Email Pipeline",
          "Mail Gateway - Templates"
        ],
        "summary": "Render Jinja2 template preview with test variables",
        "description": "Dynamically compile and preview rendered email subject and HTML using provided sample data.",
        "operationId": "render_mail_template_preview_api_v1_mail_gateway_templates_render_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailTemplateRenderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailTemplateRenderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/mail_gateway/send": {
      "post": {
        "tags": [
          "Mail Gateway & Async Email Pipeline",
          "Mail Gateway - Pipeline"
        ],
        "summary": "Enqueue and dispatch outbound email",
        "description": "Enqueue an outbound email with Jinja2 rendering and asynchronous Celery background transmission.",
        "operationId": "send_mail_api_v1_mail_gateway_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendMailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/mail_gateway/queue/{queue_id}": {
      "get": {
        "tags": [
          "Mail Gateway & Async Email Pipeline",
          "Mail Gateway - Pipeline"
        ],
        "summary": "Get mail queue delivery status",
        "description": "Retrieve detailed telemetry and delivery status for an enqueued email.",
        "operationId": "get_mail_queue_item_api_v1_mail_gateway_queue__queue_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "queue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Queue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailQueueRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectionTestResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "ConnectionTestResponse",
        "example": {
          "latency_ms": 42.5,
          "message": "SMTP handshake and authentication successful",
          "success": true
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SendMailResponse": {
        "properties": {
          "queue_id": {
            "type": "string",
            "format": "uuid",
            "title": "Queue Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "queue_id",
          "state",
          "message"
        ],
        "title": "SendMailResponse",
        "example": {
          "message": "Email dispatched successfully",
          "queue_id": "q0000000-0000-0000-0000-000000000001",
          "state": "sent"
        }
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/identity_rbac/auth/login"
          }
        }
      }
    }
  }
}